requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cx_Freeze==7.2.1
cx_Logging==3.2.0
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
lief==0.15.1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller-hooks-contrib==2024.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-docx==1.1.2
pywin32-ctypes==0.2.3
regex==2024.9.11
requests==2.32.3
setuptools==74.1.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
ttkbootstrap==1.10.1
typing_extensions==4.12.2
urllib3==2.2.3
